Cache: Ensure add/put return a promise for read-only master stores
Note that for simplicity's sake, this fix presumes that the master store
adheres to the documented behavior of the add and put APIs, i.e. that it
returns an object; if it returns a falsy value, this fix will end up
returning the result of add/put on the caching store in that case too.

This commit also updates RequestMemory tests to remove when calls and test
Cache APIs for promises.  (Without the fixes in Cache.js, the add/put tests
will fail.)