Convert QueryResults to TypeScript
Also replace references to QueryResults with
QueryResults.default.

Two test cases needed to be updated, as they were relying on
the fact that dojo/when executed a callback immediately in
the case that the provided value was synchronous. Now the
forEach function provided by QueryResults uses a Promise and
is always asynchronous.