Memory store putSync performance improvement
The existing logic splices out a modified element and splices it back in
any time an update is made with putSync. With large arrays the performance
of Array#splice is bad. Additionally, the store's index is rebuilt to account
for moved items.

This PR adds a check to see if the item being put has moved. If not, the array
is modified in-place with no splicing, and the store's index is not rebuilt.

Some boundary tests for putSync are also added to the test suite.